moved session hotkeys hook
